Change Language and status bar
I changed the way I managed the language. Now, in the file
WolfWriterLanguagesRules there is all the rules that are needed in every
languages. The WWLanguage classes have now a list of theses rules and in
cheacking in the order of the list.
Now every rule has it own description that is displayed in a brand new
status bar of WWMainWindow. Every time the software makes a correction,
a short explainaition in shown in the status bar to explain the nature
of the correction.
This satus bar also display information when opening, saving, saving as,
exporting or saving an archive. The message is display in the status bar
via the function WWMainWindow.changeMessageStatusBar .