Enhanced exportation
Now there is a better way to export books. The new file WolfWriterExport
contains WWExport classes (for now there is only 2 for txt export and
html export). Each of these classes contains options which are put in
the structure.output options. They are used by the arboressence of the
book during the exportation. Not that some of these options are python
code which are evalutated at some point in the exportation. I also
rename the "txt_output" function in "output".