Put the english rules
Extend the rules of typography for the English Language. Now, an
attribute in the structure of the book contains the languge in which is
written the book. The WWTextEdit and WWLineEdit takes this information
from the WWStructure to choose the good language. A new constants has
been made : DFT_WRITING_LANGUAGE. It is stating by default what language
should we choose when creating a new book.
Some things to do : changing the Language in the metadata should change
the language of the WWSceneEdit directly
I also added an action in the WWMainWindw that check the typography of a
book.