Ctrl-Z and Typo and new scene ; Rule conflict
Ctrl-Z and Typo, now when there is a typography correction, making undo
just after will remove both the correction and the previo modification.
Ctrl-Z and new scene : introducing the new scene is not uncluded in the
history of the document.
Rule conflict : change the conflict of Rule in French with the dialog
dash and the unbreakable space
Pasting : remove all the tabulation and replace them by space